refactor(analysis): consolidate analysis model files into a single module
