--- src/analysis/file_role.rs
@@ -1,11 +1,0 @@
-use serde::Serialize;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-pub enum FileRole {
-    Production,
-    Test,
-    Generated,
-    Config,
-    Documentation,
-    Fixture,
-}
