+++ src/analysis/mod.rs
@@ -1,11 +1,5 @@
-pub mod context;
-pub mod file_role;
-pub mod language_family;
-pub mod module_kind;
+mod model;
 pub mod path_classifier;
 
-pub use context::ArchitectureContext;
-pub use file_role::FileRole;
-pub use language_family::LanguageFamily;
-pub use module_kind::ModuleKind;
+pub use model::{ArchitectureContext, FileRole, LanguageFamily, ModuleKind};
 pub use path_classifier::{ArchitectureClassifier, classify_file_architecture};
