--- src/analysis/module_kind.rs
@@ -1,12 +1,0 @@
-use serde::Serialize;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-pub enum ModuleKind {
-    Feature,
-    Shared,
-    Infrastructure,
-    Domain,
-    Ui,
-    Cli,
-    Unknown,
-}
